Refeactored framework deployment utils
This is the continuation of the refactoring of framework/deployment_utils.go
into framework/deployment.
