+++ test/e2e/framework/BUILD
@@ -7,7 +7,6 @@ go_library(
     srcs = [
         "cleanup.go",
         "create.go",
-        "deployment_util.go",
         "exec_util.go",
         "flake_reporting_util.go",
         "framework.go",
@@ -44,7 +43,6 @@ go_library(
         "//pkg/apis/storage/v1/util:go_default_library",
         "//pkg/client/conditions:go_default_library",
         "//pkg/controller:go_default_library",
-        "//pkg/controller/deployment/util:go_default_library",
         "//pkg/controller/nodelifecycle:go_default_library",
         "//pkg/controller/service:go_default_library",
         "//pkg/features:go_default_library",
